+++ pyomo/contrib/asl_functions_lib/plugins.py
@@ -10,8 +10,8 @@
 #  ___________________________________________________________________________
 
 from pyomo.common.extensions import ExtensionBuilderFactory
-from pyomo.contrib.libaslfunctions.build import LibASLFunctionsBuilder
+from pyomo.contrib.asl_functions_lib.build import LibASLFunctionsBuilder
 
 
 def load():
-    ExtensionBuilderFactory.register("libaslfunctions")(LibASLFunctionsBuilder)
+    ExtensionBuilderFactory.register("asl_functions_lib")(LibASLFunctionsBuilder)
