+++ pyomo/contrib/asl_functions_lib/src/functions.cpp
