+++ pyomo/environ/tests/test_package_layout.py
@@ -39,7 +39,7 @@ _NON_MODULE_DIRS = {
     join('contrib', 'ampl_function_demo', 'src'),
     join('contrib', 'appsi', 'cmodel', 'src'),
     join('contrib', 'cspline_external', 'src'),
-    join('contrib', 'libaslfunctions', 'src'),
+    join('contrib', 'asl_functions_lib', 'src'),
     join('contrib', 'simplification', 'ginac', 'src'),
     join('contrib', 'pynumero', 'src'),
     join('core', 'tests', 'data', 'baselines'),
